Move BCID handling to separate file
